pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "numpy>=2.3.1",
    "openai>=1.97.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "playwright>=1.53.0",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.47.0",
    "trafilatura>=2.0.0",
]
